+++ rest_framework_services/views/query/selector_list_view.py
@@ -10,15 +10,12 @@ from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.serializers import BaseSerializer
 
+from rest_framework_services.dispatch.renderable_serializer_class import renderable_serializer_class
 from rest_framework_services.selectors.utils import dispatch_selector_for_spec
 from rest_framework_services.types.selector_kind import SelectorKind
 from rest_framework_services.types.selector_spec import SelectorSpec
 from rest_framework_services.views.spec_validation import validate_selector_view_spec
-from rest_framework_services.views.utils import (
-    get_class_attr,
-    layer_serializer_context,
-    renderable_serializer_class,
-)
+from rest_framework_services.views.utils import get_class_attr, layer_serializer_context
 
 
 class SelectorListView(ListModelMixin, GenericAPIView):